Start seperation of capture backend code into library
